feat: Doxygen comments for PeripheryController and ServerConnector
