+++ kos/navigation_system/include/navigation_system.h
@@ -7,7 +7,7 @@
  *
  * \~Russian
  * \brief Объявление методов для работы с местоположением дрона.
- * \details В файле объявлены публичныe методы, необходимые
+ * \details В файле объявлены публичныe методы, необходимые для
  * получения, хранения, обновления и передачи информации о текущем
  * местоположении дрона.
  */
