+++ kos/periphery_controller/src/periphery_controller_real.cpp
@@ -1,3 +1,16 @@
+/**
+ * \file
+ * \~English
+ * \brief Implementation of methods for hardware peripherals control.
+ * \details The file contains implementation of methods,
+ * that control peripheral drone devices via GPIO interface.
+ *
+ * \~Russian
+ * \brief Реализация методов для работы с аппаратной периферией.
+ * \details В файле реализованы методы, управляющие периферийными устройствами
+ * дрона через интерфейс GPIO.
+ */
+
 #include "../include/periphery_controller.h"
 
 #include <coresrv/hal/hal_api.h>
@@ -12,6 +25,7 @@
 #define NK_USE_UNQUALIFIED_NAMES
 #include <drone_controller/PeripheryController.edl.h>
 
+/** \cond */
 #define NAME_MAX_LENGTH 64
 
 char gpio[] = "gpio0";
@@ -24,7 +38,18 @@ uint8_t pinKillSwitchFirst = 22;
 uint8_t pinKillSwitchSecond = 27;
 
 bool killSwitchEnabled;
+/** \endcond */
 
+/**
+ * \~English Sets the mode of specified pin power supply.
+ * \param[in] pin Pin to set mode.
+ * \param[in] mode Mode. 1 is high, 0 is low.
+ * \return Returns 1 on successful mode set, 0 otherwise.
+ * \~Russian Устанавливает режим подачи энергии на заданный пин.
+ * \param[in] pin Пин, для которого устанавливается режим.
+ * \param[in] mode Режим. 1 -- высокий, 0 -- низкий.
+ * \return Возвращает 1, если режим был успешно установлен, иначе -- 0.
+ */
 int setPin(uint8_t pin, bool mode) {
     Retcode rc = GpioOut(gpioHandler, pin, mode);
     if (rcOk != rc) {
